Final Version
Comments in code. The function lstsqu_slopes is used to create the figures for part 2. Everything else pertains to part 1. Part 1A generates a computationally taxing matrix depending on the input, therefore it is own script that saves its variable into innie.mat. That is loaded onto Part_1B. Part_1A and Part_1B could be used for spreadsheets that follow the same convention and name as biomart and protein_consensus but can have more any amount of rows.